Add MonoDevelop Addin
This addin integrates AutoTest.NET in MonoDevelop. The status icon allows
to enable or disable watching for changes. When enabled AutoTest.NET
will watch for changes in the current solution directory and any sub-
directories, kick off a build when it finds a change and then run the
unit tests.

The buttons on the left allow filtering the build and test messages. The
button at the very left controls optimistic builds: when enabled only
the project which contains the changed file will be built, otherwise the
entire solution.

A double click on a build message places the cursor in the failing line
of the source file.

Selecting a failing test in the list of messages allows to run or debug
this single test.